Refactor: Denesting moved to its own module
